Bugfixes
- added return for Grid::Remove
- changed size to double

+++ include/particles.hpp
@@ -23,7 +23,7 @@ public:
 
 public:
     std::vector<Particle> particles;
-    int size;
+    double size;
 
 private:
 private:
